debt - cleanup config migration
